backend/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
azure-ai-documentintelligence
openai
pydantic
pydantic-settings
pillow
python-multipart
httpx
pytest
pytest-asyncio
aiohttp
pdfplumber
python-docx
reportlab
azure-communication-email
